+++ src/main/java/ru/yandex/practicum/filmorate/storage/database/UserDbStorage.java
@@ -4,8 +4,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.dao.DataAccessException;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.PreparedStatementCreator;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import org.springframework.jdbc.core.namedparam.SqlParameterSource;
+import org.springframework.jdbc.core.namedparam.SqlParameterSourceUtils;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
@@ -17,31 +19,30 @@ import ru.yandex.practicum.filmorate.storage.database.mappers.user.FriendsRowMap
 import ru.yandex.practicum.filmorate.storage.database.mappers.user.ResponseUserRowMapper;
 import ru.yandex.practicum.filmorate.storage.database.response.ResponseUser;
 
-import java.sql.Date;
-import java.sql.PreparedStatement;
 import java.util.List;
+import java.util.Objects;
 
 @Slf4j
 @Repository
 @Qualifier("userDbStorage")
 @RequiredArgsConstructor
 public class UserDbStorage implements UserStorage {
-    private final JdbcTemplate jdbcTemplate;
+    private final NamedParameterJdbcTemplate jdbc;
     private final ResponseUserRowMapper mapper;
     private final FriendsRowMapper friendMapper;
 
     @Override
     public ResponseUser createUser(RequestUser request) {
-        String sqlQuery = "INSERT INTO users (email,login,name,birthday) VALUES (?,?,?,?);";
+        SqlParameterSource[] batch = SqlParameterSourceUtils.createBatch(request.getUserDto());
+        String sqlQuery = "INSERT INTO users (email,login,name,birthday) VALUES (:email,:login,:name,:birthday)";
         KeyHolder keyHolder = new GeneratedKeyHolder();
-        PreparedStatementCreator psc = getPreparedStatement(request, sqlQuery);
-        int result;
-        result = jdbcTemplate.update(psc, keyHolder);
-        if (result == 0) {
+        int[] result;
+        result = jdbc.batchUpdate(sqlQuery, batch, keyHolder);
+        if (result.length == 0) {
             throw new ErrorAddingData("пользователь не был добавлен");
         }
         Integer userId = (Integer) keyHolder.getKey();
-        if (userId != null) {
+        if (Objects.nonNull(userId)) {
             return getUserById(userId);
         } else {
             throw new ErrorAddingData("из БД не был возвращен ID нового пользователя");
@@ -50,15 +51,18 @@ public class UserDbStorage implements UserStorage {
 
     @Override
     public ResponseUser updateUser(RequestUser request) {
-        String sqlQuery = "UPDATE users SET email = ?,login = ?, name = ?,birthday = ? WHERE user_id = ?;";
-        PreparedStatementCreator psc = getPreparedStatement(request, sqlQuery);
-        int result = 0;
+        SqlParameterSource[] batch = SqlParameterSourceUtils.createBatch(request.getUserDto());
+        String sqlQuery = "UPDATE users " +
+                "SET email = :email, login = :login, name = :name, birthday = :birthday " +
+                "WHERE user_id = " + request.getId();
+        int[] result;
         try {
-            result = jdbcTemplate.update(psc);
+            result = jdbc.batchUpdate(sqlQuery, batch);
         } catch (DataAccessException e) {
             log.debug("ошибка при обращении к БД: {}", e.getMessage());
+            throw new ErrorAddingData(e.getMessage());
         }
-        if (result == 0) {
+        if (result.length == 0) {
             throw new ErrorAddingData("данные о пользователе не были изменены");
         }
         return getUserById(request.getId());
@@ -66,15 +70,17 @@ public class UserDbStorage implements UserStorage {
 
     @Override
     public void deleteUser(int userId) {
-        String sqlQuery = "DELETE FROM users WHERE user_id = ?;";
-        jdbcTemplate.update(sqlQuery, userId);
+        SqlParameterSource namedParameters = new MapSqlParameterSource().addValue("userID", userId);
+        String sqlQuery = "DELETE FROM users WHERE user_id = :userID";
+        jdbc.update(sqlQuery, namedParameters);
     }
 
     @Override
     public ResponseUser getUserById(int userID) {
-        String sqlQuery = "SELECT * FROM users WHERE user_id = ?;";
+        SqlParameterSource namedParameters = new MapSqlParameterSource().addValue("userID", userID);
+        String sqlQuery = "SELECT * FROM users WHERE user_id = :userID";
         try {
-            return jdbcTemplate.queryForObject(sqlQuery, mapper, userID);
+            return jdbc.queryForObject(sqlQuery, namedParameters, mapper);
         } catch (DataAccessException e) {
             throw new NotFoundUserException();
         }
@@ -83,25 +89,28 @@ public class UserDbStorage implements UserStorage {
     @Override
     public List<ResponseUser> getAllUsers() {
         String sqlQuery = "SELECT * FROM users;";
-        return jdbcTemplate.query(sqlQuery, mapper);
+        return jdbc.query(sqlQuery, mapper);
     }
 
     @Override
     public void addFriend(int userID, int friendID) {
-        String sqlQuery = "INSERT INTO friends (user_id,friend_id) VALUES (?,?);";
+        SqlParameterSource namedParameters = new MapSqlParameterSource()
+                .addValue("userID", userID).addValue("friendID", friendID);
+        String sqlQuery = "INSERT INTO friends (user_id,friend_id) VALUES (:userID,:friendID);";
         try {
-            jdbcTemplate.update(sqlQuery, userID, friendID);
+            jdbc.update(sqlQuery, namedParameters);
         } catch (DataAccessException e) {
             log.debug("ошибка при добавлении пользователя в друзья: {}", e.getMessage());
         }
-
     }
 
     @Override
     public void deleteFriend(int userID, int friendID) {
-        String sqlQuery = "DELETE FROM friends WHERE user_id = ?;";
+        SqlParameterSource namedParameters = new MapSqlParameterSource()
+                .addValue("userID", userID).addValue("friendID", friendID);
+        String sqlQuery = "DELETE FROM friends WHERE user_id = :userID AND :friendID";
         try {
-            jdbcTemplate.update(sqlQuery, userID);
+            jdbc.update(sqlQuery, namedParameters);
         } catch (DataAccessException e) {
             log.debug("ошибка при удалении друзей: {}", e.getMessage());
         }
@@ -109,22 +118,9 @@ public class UserDbStorage implements UserStorage {
 
     @Override
     public List<Integer> getListFriends(Integer userid) {
-        String sklQuery = "SELECT friend_id FROM friends WHERE user_id = ?;";
-        return jdbcTemplate.query(sklQuery, friendMapper, userid);
-    }
-
-
-    private PreparedStatementCreator getPreparedStatement(RequestUser request, String sqlQuery) {
-        return con -> {
-            PreparedStatement stm = con.prepareStatement(sqlQuery, new String[]{"user_id"});
-            stm.setString(1, request.getEmail());
-            stm.setString(2, request.getLogin());
-            stm.setString(3, request.getName());
-            stm.setDate(4, Date.valueOf(request.getBirthday()));
-            if (request.getId() != null) {
-                stm.setInt(5, request.getId());
-            }
-            return stm;
-        };
+        SqlParameterSource namedParameters = new MapSqlParameterSource()
+                .addValue("userID", userid);
+        String sklQuery = "SELECT friend_id FROM friends WHERE user_id = :userID";
+        return jdbc.query(sklQuery,namedParameters, friendMapper);
     }
 }
